+++ certbot-apache/certbot_apache/tests/display_ops_test.py
@@ -1,4 +1,4 @@
-"""Test certbot_apache.display_ops."""
+"""Test certbot_apache._internal.display_ops."""
 import unittest
 
 import mock
@@ -11,12 +11,12 @@ from certbot.tests import util as certbot_util
 
 from certbot_apache import obj
 
-from certbot_apache.display_ops import select_vhost_multiple
+from certbot_apache._internal.display_ops import select_vhost_multiple
 from certbot_apache.tests import util
 
 
 class SelectVhostMultiTest(unittest.TestCase):
-    """Tests for certbot_apache.display_ops.select_vhost_multiple."""
+    """Tests for certbot_apache._internal.display_ops.select_vhost_multiple."""
 
     def setUp(self):
         self.base_dir = "/example_path"
@@ -45,7 +45,7 @@ class SelectVhostMultiTest(unittest.TestCase):
         self.assertFalse(vhs)
 
 class SelectVhostTest(unittest.TestCase):
-    """Tests for certbot_apache.display_ops.select_vhost."""
+    """Tests for certbot_apache._internal.display_ops.select_vhost."""
 
     def setUp(self):
         self.base_dir = "/example_path"
@@ -54,7 +54,7 @@ class SelectVhostTest(unittest.TestCase):
 
     @classmethod
     def _call(cls, vhosts):
-        from certbot_apache.display_ops import select_vhost
+        from certbot_apache._internal.display_ops import select_vhost
         return select_vhost("example.com", vhosts)
 
     @certbot_util.patch_get_utility()
@@ -81,9 +81,9 @@ class SelectVhostTest(unittest.TestCase):
     def test_no_vhosts(self):
         self.assertEqual(self._call([]), None)
 
-    @mock.patch("certbot_apache.display_ops.display_util")
+    @mock.patch("certbot_apache._internal.display_ops.display_util")
     @certbot_util.patch_get_utility()
-    @mock.patch("certbot_apache.display_ops.logger")
+    @mock.patch("certbot_apache._internal.display_ops.logger")
     def test_small_display(self, mock_logger, mock_util, mock_display_util):
         mock_display_util.WIDTH = 20
         mock_util().menu.return_value = (display_util.OK, 0)
