+++ certbot-apache/docs/api/display_ops.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_apache.display_ops`
+:mod:`certbot_apache._internal.display_ops`
 -------------------------------------
 
-.. automodule:: certbot_apache.display_ops
+.. automodule:: certbot_apache._internal.display_ops
    :members:
