+++ certbot-nginx/certbot_nginx/tests/display_ops_test.py
@@ -1,4 +1,4 @@
-"""Test certbot_apache.display_ops."""
+"""Test certbot_apache._internal.display_ops."""
 import unittest
 
 from certbot.display import util as display_util
